requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
pydantic[email]==2.9.2
httpx==0.27.2          # For testing
pytest==8.3.3          # For testing
pytest-asyncio==0.24.0 # For testing
